+++ packages/airless-core/airless/core/operator/redirect.py
@@ -1,7 +1,7 @@
 
 from airless.core.operator import BaseEventOperator
 
-from airless.core.config import get_config
+from airless.core.utils import get_config
 
 
 class RedirectOperator(BaseEventOperator):