+++ packages/airless-google-cloud-core/airless/google/cloud/core/operator/error.py
@@ -4,7 +4,7 @@ import time
 
 from datetime import datetime
 
-from airless.core.config import get_config
+from airless.core.utils import get_config
 from airless.core.dto import BaseDto
 from airless.google.cloud.core.operator import GoogleBaseEventOperator