+++ packages/airless-google-cloud-core/airless/google/cloud/pubsub/hook/pubsub.py
@@ -4,7 +4,7 @@ import json
 from google.cloud import pubsub_v1
 
 from airless.core.hook import QueueHook
-from airless.core.config import get_config
+from airless.core.utils import get_config
 
 
 class GooglePubsubHook(QueueHook):