+++ packages/airless-google-cloud-storage/airless/google/cloud/storage/hook/storage.py
@@ -9,7 +9,7 @@ from google.cloud import storage
 from google.cloud.storage.retry import DEFAULT_RETRY
 from pyarrow import fs, parquet
 
-from airless.core.config import get_config
+from airless.core.utils import get_config
 from airless.core.hook import BaseHook, FileHook