+++ packages/airless-captcha/CHANGELOG.md
@@ -1,5 +1,6 @@
 
 **unreleased**
+- [Feature] Enhance dependencies
 
 **v0.0.4**
 - [Feature] Change package to mono repo