+++ packages/airless-google-cloud-secret-manager/CHANGELOG.md
@@ -1,5 +1,7 @@
 
 **unreleased**
+- [Feature] Get method `get_config` to new path
+- [Feature] Enhance dependencies
 
 **v0.0.1**
 - [Feature] Package created